lib/mpi: avoid null pointer deref in mpi_cmp_ui()
During NVMeTCP Authentication a controller can trigger a kernel
oops by specifying the 8192 bit Diffie Hellman group and passing
a correctly sized, but zeroed buffer. mpi_cmp_ui() was detecting
this if the second parameter was 0, but 1 is passed from
dh_is_pubkey_valid(). This causes the null pointer u->d to be
dereferenced towards the end of mpi_cmp_ui()

Signed-off-by: Mark O'Donovan <shiftee@posteo.net>